+++ Inventory.h
@@ -32,12 +32,12 @@ class Inventory
       std::string myItemDesc;
       int myCategoryID;
       int myMerchantID;
-      double myPrice;
+      int myPrice;
       int myQuantity;
 
    public:
       Inventory();
-      Inventory(const int sku, const std::string& itemDesc, const int categoryID, const int merchantID, const double price, const int quantity)
+      Inventory(const int sku, const std::string& itemDesc, const int categoryID, const int merchantID, const int price, const int quantity)
 	  : mySKU(sku), myItemDesc(itemDesc), myCategoryID(categoryID), myMerchantID(merchantID), myPrice(price), myQuantity(quantity) {}
       Inventory(const Inventory& inventory);
       Inventory(const YAML::Node& node);
@@ -49,14 +49,14 @@ class Inventory
       const std::string& getItemDesc() const;
       int getCategoryID() const;
       int getMerchantID() const;
-      double getPrice() const;
+      int getPrice() const;
       int getQuantity() const;
 
       void setSKU(const int sku);
       void setItemDesc(const std::string& itemDesc);
       void setCategoryID(int categoryID);
       void setMerchantID(int merchantID);
-      void setPrice(const double price);
+      void setPrice(const int price);
       void setQuantity(const int quantity); 
 };
